fixed again

+++ django/app/models.py
@@ -27,7 +27,7 @@ class JobPost(models.Model):
     description = models.TextField()
     company = models.CharField(max_length=255)
     location = models.CharField(max_length=255)
-    salary = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
+    salary = models.FloatField(null=True, blank=True)
     posted_at = models.DateTimeField(auto_now_add=True)
     is_active = models.BooleanField(default=True)
